Implement wing animations and head animation for crows. Move TreasureChestModel to model package. Remove newlines in multiple biome classes.
