+++ src/main/java/com/fabriccommunity/spookytime/client/model/CrowEntityModel.java
@@ -1,10 +1,9 @@
 package com.fabriccommunity.spookytime.client.model;
 
+import com.fabriccommunity.spookytime.entity.CrowEntity;
 import net.minecraft.client.model.Cuboid;
 import net.minecraft.client.render.entity.model.EntityModel;
 import net.minecraft.util.math.MathHelper;
-
-import com.fabriccommunity.spookytime.entity.CrowEntity;
 
 public class CrowEntityModel extends EntityModel<CrowEntity> {
 	private final Cuboid body;
@@ -75,15 +74,25 @@ public class CrowEntityModel extends EntityModel<CrowEntity> {
 		cuboid.yaw = y;
 		cuboid.roll = z;
 	}
-	
+
+	@Override
+	public void animateModel(CrowEntity entity_1, float float_1, float float_2, float float_3) {
+		super.animateModel(entity_1, float_1, float_2, float_3);
+	}
+
 	@Override
 	public void setAngles(CrowEntity crow, float f, float f1, float f2, float f3, float f4, float f5) {
 		super.setAngles(crow, f, f1, f2, f3, f4, f5);
+
+		head.pitch = f4 * 0.017453292F;
+		head.yaw = f3 * 0.017453292F;
+
 		if (crow.isInAir()) {
 			leftLeg.pitch = 0.55F;
 			rightLeg.pitch = 0.55F;
-			
-			//make wings flap here when you figure out how
+
+			this.leftWing.roll = 0.0873F + f2;
+			this.rightWing.roll = -0.0873F - f2;
 		} else {
 			leftLeg.pitch = -0.6109F + MathHelper.cos(f * 0.6662F) * 1.4F * f1;
 			rightLeg.pitch = -0.6109F + MathHelper.cos(f * 0.6662F + 3.1415927F) * 1.4F * f1;
