+++ src/main/java/com/fabriccommunity/spookytime/world/biome/SpookySeaBiome.java
@@ -18,7 +18,6 @@ import com.fabriccommunity.spookytime.world.feature.SpookyBiomeFeatures;
 
 // TODO
 public class SpookySeaBiome extends SpookyBaseBiome {
-	
 	protected static final int GRASS_COLOR = 0x20003B;
 	protected static final int FOLIAGE_COLOR = 0x20003B;
