+++ src/main/java/com/fabriccommunity/spookytime/world/biome/SpookyShoreBiome.java
@@ -21,7 +21,6 @@ import com.fabriccommunity.spookytime.world.feature.SpookyBiomeFeatures;
 
 // TODO
 public class SpookyShoreBiome extends SpookyBaseBiome {
-	
 	protected static final int GRASS_COLOR = 0x20003B;
 	protected static final int FOLIAGE_COLOR = 0x20003B;
