+++ src/main/java/com/fabriccommunity/spookytime/world/biome/SpookySwampBiome.java
@@ -22,7 +22,6 @@ import com.fabriccommunity.spookytime.world.feature.SpookyBiomeFeatures;
 
 // TODO
 public class SpookySwampBiome extends SpookyBaseBiome {
-	
 	protected static final int GRASS_COLOR = 0x20003B;
 	protected static final int FOLIAGE_COLOR = 0x20003B;
